+++ timer.cpp
@@ -2,11 +2,12 @@
 
 #include <cmath>
 
-bool timer::PeriodicTimer::tick(float dt) {
+std::optional<float> timer::PeriodicTimer::tick(float dt) {
     accumulated += dt;
     if (accumulated >= interval) {
+        auto result{accumulated};
         accumulated = std::fmod(accumulated, interval);
-        return true;
+        return result;
     }
-    return false;
+    return {};
 }