+++ timer.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <optional>
+
 namespace timer {
 
 struct PeriodicTimer {
     const float interval;
     float accumulated = 0;
 
-    bool tick(float dt);
+    std::optional<float> tick(float dt);
 };
 
 }  // namespace timer